drivers: iicps: Fix MISRAC 10.3 rule violation
Fixed the following violation:
misra_c_2012_rule_10_3_violation: Implicit conversion from
essential type "signed 64-bit int" to different or narrower
essential type "signed 32-bit int"


Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/iicps/src/xiicps_master.c
@@ -393,7 +393,7 @@ s32 XIicPs_MasterSendPolled(XIicPs *InstancePtr, u8 *MsgPtr,
 
 	/* Set the Status for XST_SUCCESS */
 	if (((IntrStatusReg & Intrs) == 0U) && (timeout != TX_MAX_LOOPCNT)) {
-		Status = XST_SUCCESS;
+		Status = (s32)XST_SUCCESS;
 	}
 
 	return Status;
